converting onboarding controller to typescript
